refactor: reduce sprawl
