--- containers/ecr-viewer/src/app/data/db/utils/types.ts
@@ -1,5 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type AnyDb = any;
-
-export type DialectType = "sqlserver" | "postgres";
-export type SchemaType = "core" | "extended";
