+++ containers/ecr-viewer/src/app/services/listEcrDataService.ts
@@ -5,7 +5,7 @@ import { getSql } from "@/app/api/services/dialects/common";
 import { Common, ecr_data } from "@/app/api/services/types/common";
 import { Core } from "@/app/api/services/types/core";
 import { Extended } from "@/app/api/services/types/extended";
-import { dbSchema } from "@/app/data/db/utils/db-config";
+import { dbSchema } from "@/app/api/services/utils/db-config";
 import { DateRangePeriod } from "@/app/utils/date-utils";
 
 import { formatDate, formatDateTime } from "./formatDateService";
